Decent stab at newPosition(). Two failures in edge cases:
/Users/sas/Projects/CoPilotPlugin/CoPilotPluginTests/DiffTests.swift:168: error: -[CoPilotPluginTests.DiffTests test_adjustPos] : failed - expected to equal <52>, got <51>

/Users/sas/Projects/CoPilotPlugin/CoPilotPluginTests/DiffTests.swift:267: error: -[CoPilotPluginTests.DiffTests test_preserve_position] : failed - expected to equal <27>, got <28>

I believe problem is that a delete diff can pull the posPointer back such that the next iteration does not satisfy

Diff.swift:138            if diffPointer <= posPointer {

which then leads to the subsequent insert at the same position not being applied. There needs to be some kind of delay in applying the adjustment when the diff was a del (or ins?). Or the comparison in 138 is only applied after a .DiffEqual operation?

Note that using “if diffPointer < posPointer” also has two but different failures:

/Users/sas/Projects/CoPilotPlugin/CoPilotPluginTests/DiffTests.swift:169: error: -[CoPilotPluginTests.DiffTests test_adjustPos] : failed - expected to equal <53>, got <52>

/Users/sas/Projects/CoPilotPlugin/CoPilotPluginTests/DiffTests.swift:184: error: -[CoPilotPluginTests.DiffTests test_adjustPos] : failed - expected to equal <88>, got <87>