Fixes #14
The issue was that the underlying _document was updated on each update() but the command was only sent when the buffer throttle time was exceeded between calls.

This leads to the diff’s baseRev being unequal and the patch being rejected. For the client this means resetting to the server status (rejecting all local changes). For the server this means an unnecessary round trip from the client requesting a reset.