WIP trying to fix `test_sendChanges`.
There are issues with race conditions here and the whole communication needs a proper re-design. The issue is that the original bonjour-based set up was not built for a star-topology where the server isn't the hub. The RabbitServer version is different, the master document isn't also the hub that the clients connect to. Not sure it's really worth trying to hack this different topology in. It may be better to revisit the communication protocol and re-design it with this more general topology in mind. Bonjour connectivity would then just be a simpler version of this.